+++ option/tun.go
@@ -11,6 +11,8 @@ type TunInboundOptions struct {
 	ExcludeUID             Listable[uint32] `json:"exclude_uid,omitempty"`
 	ExcludeUIDRange        Listable[string] `json:"exclude_uid_range,omitempty"`
 	IncludeAndroidUser     Listable[int]    `json:"android_user,omitempty"`
+	IncludePackage         Listable[string] `json:"include_package,omitempty"`
+	ExcludePackage         Listable[string] `json:"exclude_package,omitempty"`
 	EndpointIndependentNat bool             `json:"endpoint_independent_nat,omitempty"`
 	UDPTimeout             int64            `json:"udp_timeout,omitempty"`
 	Stack                  string           `json:"stack,omitempty"`